Break down SpawnEditorWorkspace into smaller files
Extract 14 components and hooks from SpawnEditorWorkspace.tsx

- SpawnEditorDialogs: 5 confirmation dialogs
- SpawnSettingsHeader, BasicDetailsCard, MetadataCard: UI sections
- TriggerTypeSelector and 6 trigger config components
- useAssetDraftCache hook and triggerConfigHelpers

Also updates Headless UI to use named exports (DisclosureButton,
ComboboxInput) instead of compound syntax, and adds ESLint rule
to warn on deprecated APIs.
