+++ eslint.config.js
@@ -12,6 +12,13 @@ export default tseslint.config(
     languageOptions: {
       ecmaVersion: 2020,
       globals: globals.browser,
+      parserOptions: {
+        project: [
+          "./tsconfig.app.json",
+          "./tsconfig.node.json",
+          "./tsconfig.test.json",
+        ],
+      },
     },
     plugins: {
       "react-hooks": reactHooks,
@@ -23,6 +30,7 @@ export default tseslint.config(
         "warn",
         { allowConstantExport: true },
       ],
+      "@typescript-eslint/no-deprecated": "warn",
     },
   },
 );
