+++ src/components/configuration/SpawnEditorWorkspace.tsx
@@ -6,73 +6,15 @@ import React, {
   memo,
   useCallback,
 } from "react";
-import { useStreamerbotCommands } from "../../hooks/useStreamerbotCommands";
-import { HUICombobox } from "../common";
-import { Button } from "../ui/Button";
-import { Card, CardHeader, CardTitle, CardContent } from "../ui/Card";
-import { Input } from "../ui/Input";
-import { Switch } from "../ui/Switch";
 import { StreamerbotService } from "../../services/streamerbotService";
 import { toast } from "sonner";
-
-const SBCommandAliasCombobox: React.FC<{
-  value: string;
-  onChange: (v: string) => void;
-}> = ({ value, onChange }) => {
-  const { commands, loading, refresh } = useStreamerbotCommands();
-  const options = React.useMemo(() => {
-    const seen = new Set<string>();
-    const out: { value: string; label: string }[] = [];
-    (commands || []).forEach((cmd) => {
-      (cmd.commands || []).forEach((a) => {
-        const key = a.toLowerCase();
-        if (!seen.has(key) && a.trim()) {
-          seen.add(key);
-          out.push({ value: a, label: cmd.name });
-        }
-      });
-    });
-    out.sort((a, b) => a.value.localeCompare(b.value));
-    return out;
-  }, [commands]);
-
-  return (
-    <div className="flex items-center gap-2 w-full">
-      <div className="flex-1">
-        <HUICombobox
-          value={value}
-          onChange={onChange}
-          onSelect={onChange}
-          options={options}
-          isLoading={loading}
-          placeholder="Enter command alias (e.g., scene1, alert)"
-        />
-      </div>
-      <button
-        type="button"
-        onClick={() => refresh()}
-        disabled={loading}
-        className="text-xs text-[rgb(var(--color-accent))] hover:text-[rgb(var(--color-accent-hover))] disabled:opacity-50"
-      >
-        Refresh
-      </button>
-    </div>
-  );
-};
-import moment from "moment-timezone/builds/moment-timezone-with-data-1970-2030";
 import { usePanelState } from "../../hooks/useLayout";
 import { SpawnService } from "../../services/spawnService";
 import type { Spawn } from "../../types/spawn";
 import type { RandomizationBucket } from "../../types/spawn";
 import type { Trigger, TriggerType } from "../../types/spawn";
 import { getDefaultTrigger } from "../../types/spawn";
-import type { MediaAssetProperties } from "../../types/media";
-import { ConfirmDialog } from "../common/ConfirmDialog";
 import AssetSettingsForm from "./asset-settings/AssetSettingsForm";
-import {
-  getNextActivation,
-  formatNextActivation,
-} from "../../utils/scheduling";
 import { validateTrigger } from "../../utils/triggerValidation";
 import { RandomizationBucketsSection } from "./RandomizationBucketsSection";
 import { validateRandomizationBuckets } from "../../utils/randomizationBuckets";
@@ -82,87 +24,27 @@ import {
   useMediaSpawnerEvent,
 } from "../../hooks/useMediaSpawnerEvent";
 
-const buildTimezoneOptions = () => {
-  const now = Date.now();
-  return moment.tz
-    .names()
-    .map((zone) => {
-      const offsetMin = moment.tz(now, zone).utcOffset();
-      const sign = offsetMin >= 0 ? "+" : "-";
-      const hh = String(Math.floor(Math.abs(offsetMin) / 60)).padStart(2, "0");
-      const mm = String(Math.abs(offsetMin) % 60).padStart(2, "0");
-      return {
-        value: zone,
-        label: `(UTC${sign}${hh}:${mm}) ${zone}`,
-        offset: offsetMin,
-      };
-    })
-    .sort((a, b) => a.offset - b.offset || a.label.localeCompare(b.label));
-};
+// Extracted components
+import { SpawnEditorDialogs } from "./SpawnEditorDialogs";
+import { SpawnSettingsHeader } from "./SpawnSettingsHeader";
+import { BasicDetailsCard } from "./BasicDetailsCard";
+import { TriggerTypeSelector } from "./TriggerTypeSelector";
+import { MetadataCard } from "./MetadataCard";
 
-const timezoneOptions = buildTimezoneOptions();
-const dayOfWeekOptions = [
-  { value: 0, label: "Sunday" },
-  { value: 1, label: "Monday" },
-  { value: 2, label: "Tuesday" },
-  { value: 3, label: "Wednesday" },
-  { value: 4, label: "Thursday" },
-  { value: 5, label: "Friday" },
-  { value: 6, label: "Saturday" },
-];
+// Trigger config components
+import {
+  getCommandConfig,
+  getChannelPointConfig,
+  CommandTriggerConfig,
+  ChannelPointTriggerConfig,
+  SubscriptionTriggerConfig,
+  GiftSubTriggerConfig,
+  CheerTriggerConfig,
+  TimeTriggerConfig,
+} from "./triggers";
 
-const getCommandConfig = (trigger: Trigger | null) => {
-  if (trigger?.type === "streamerbot.command") {
-    return trigger.config;
-  }
-  return null;
-};
-
-const getChannelPointConfig = (trigger: Trigger | null) => {
-  if (trigger?.type === "twitch.channelPointReward") {
-    return trigger.config;
-  }
-  return null;
-};
-
-const getSubscriptionConfig = (trigger: Trigger | null) =>
-  trigger?.type === "twitch.subscription" ? trigger.config : null;
-const getGiftSubConfig = (trigger: Trigger | null) =>
-  trigger?.type === "twitch.giftSub" ? trigger.config : null;
-const getCheerConfig = (trigger: Trigger | null) =>
-  trigger?.type === "twitch.cheer" ? trigger.config : null;
-const getAtDateTimeConfig = (trigger: Trigger | null) =>
-  trigger?.type === "time.atDateTime" ? trigger.config : null;
-const getDailyAtConfig = (trigger: Trigger | null) =>
-  trigger?.type === "time.dailyAt" ? trigger.config : null;
-const getEveryNMinutesConfig = (trigger: Trigger | null) =>
-  trigger?.type === "time.everyNMinutes" ? trigger.config : null;
-const getMinuteOfHourConfig = (trigger: Trigger | null) =>
-  trigger?.type === "time.minuteOfHour" ? trigger.config : null;
-const getWeeklyAtConfig = (trigger: Trigger | null) => {
-  if (trigger?.type !== "time.weeklyAt") return null;
-  const config = trigger.config as {
-    daysOfWeek?: number[];
-    dayOfWeek?: number;
-    time: string;
-    timezone: string;
-  };
-  // Handle old config format (dayOfWeek) by migrating to daysOfWeek array
-  const daysOfWeek = Array.isArray(config.daysOfWeek)
-    ? config.daysOfWeek
-    : typeof config.dayOfWeek === "number"
-      ? [config.dayOfWeek]
-      : [1]; // Default to Monday if neither exists
-  return {
-    daysOfWeek,
-    time: config.time,
-    timezone: config.timezone,
-  };
-};
-const getMonthlyOnConfig = (trigger: Trigger | null) =>
-  trigger?.type === "time.monthlyOn"
-    ? (trigger.config as { dayOfMonth: number; time: string; timezone: string })
-    : null;
+// Hooks
+import { useAssetDraftCache } from "./hooks";
 
 const SpawnEditorWorkspace: React.FC = memo(() => {
   const {
@@ -210,219 +92,19 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
   const [bucketsDraft, setBucketsDraft] = useState<RandomizationBucket[]>([]);
   const [duration, setDuration] = useState<number>(0);
 
-  const assetDraftCacheRef = useRef<
-    Record<
-      string,
-      {
-        draftValues: Partial<MediaAssetProperties>;
-      }
-    >
-  >({});
+  // Use the extracted cache hook
+  const {
+    setCacheEntry,
+    getCacheEntry,
+    clearCacheEntry,
+    clearCacheEntryUnsafe,
+    clearAllCacheEntries,
+  } = useAssetDraftCache({
+    selectedSpawnId,
+    selectedSpawnAssetId,
+    centerPanelMode,
+  });
 
-  // Cache key validation and generation functions
-  const validateCacheKey = useCallback(
-    (
-      spawnId: string | undefined,
-      spawnAssetId: string | undefined,
-    ): string | null => {
-      if (!spawnId || !spawnAssetId) {
-        if (process.env.NODE_ENV === "development") {
-          console.warn(
-            "Cache key validation failed: missing spawnId or spawnAssetId",
-            { spawnId, spawnAssetId },
-          );
-        }
-        return null;
-      }
-
-      // Sanitize keys to prevent invalid characters
-      const sanitizedSpawnId = spawnId.replace(/[|]/g, "_");
-      const sanitizedSpawnAssetId = spawnAssetId.replace(/[|]/g, "_");
-
-      const key = `${sanitizedSpawnId}|${sanitizedSpawnAssetId}`;
-
-      if (process.env.NODE_ENV === "development") {
-        console.debug("Generated cache key:", key, {
-          originalSpawnId: spawnId,
-          originalSpawnAssetId: spawnAssetId,
-        });
-      }
-
-      return key;
-    },
-    [],
-  );
-
-  const generateCacheKey = useCallback(
-    (spawnId: string | undefined, spawnAssetId: string | undefined): string => {
-      const key = validateCacheKey(spawnId, spawnAssetId);
-      if (!key) {
-        throw new Error(
-          `Invalid cache key parameters: spawnId=${spawnId}, spawnAssetId=${spawnAssetId}`,
-        );
-      }
-      return key;
-    },
-    [validateCacheKey],
-  );
-
-  // Cache isolation boundary validation functions
-  const validateCacheScope = useCallback(
-    (
-      spawnId: string | undefined,
-      spawnAssetId: string | undefined,
-      operation: string,
-    ) => {
-      if (!spawnId || !spawnAssetId) {
-        const error = `Cache ${operation} failed: missing spawnId or spawnAssetId`;
-        if (process.env.NODE_ENV === "development") {
-          console.error(error, { spawnId, spawnAssetId, operation });
-        }
-        throw new Error(error);
-      }
-
-      // Validate that we're operating on the currently selected spawn
-      if (spawnId !== selectedSpawnId) {
-        const error = `Cache ${operation} boundary violation: spawnId mismatch (expected: ${selectedSpawnId}, got: ${spawnId})`;
-        if (process.env.NODE_ENV === "development") {
-          console.error(error, {
-            expectedSpawnId: selectedSpawnId,
-            actualSpawnId: spawnId,
-            spawnAssetId,
-            operation,
-          });
-        }
-        throw new Error(error);
-      }
-
-      if (process.env.NODE_ENV === "development") {
-        console.debug(`Cache ${operation} scope validated:`, {
-          spawnId,
-          spawnAssetId,
-        });
-      }
-    },
-    [selectedSpawnId],
-  );
-
-  const validateAssetScope = useCallback(
-    (spawnAssetId: string | undefined, operation: string) => {
-      if (!spawnAssetId) {
-        const error = `Asset ${operation} failed: missing spawnAssetId`;
-        if (process.env.NODE_ENV === "development") {
-          console.error(error, { spawnAssetId, operation });
-        }
-        throw new Error(error);
-      }
-
-      // Validate that we're operating on the currently selected asset (if in asset-settings mode)
-      if (
-        centerPanelMode === "asset-settings" &&
-        spawnAssetId !== selectedSpawnAssetId
-      ) {
-        const error = `Asset ${operation} boundary violation: spawnAssetId mismatch (expected: ${selectedSpawnAssetId}, got: ${spawnAssetId})`;
-        if (process.env.NODE_ENV === "development") {
-          console.error(error, {
-            expectedSpawnAssetId: selectedSpawnAssetId,
-            actualSpawnAssetId: spawnAssetId,
-            operation,
-          });
-        }
-        throw new Error(error);
-      }
-
-      if (process.env.NODE_ENV === "development") {
-        console.debug(`Asset ${operation} scope validated:`, { spawnAssetId });
-      }
-    },
-    [centerPanelMode, selectedSpawnAssetId],
-  );
-
-  // Immutable cache operations helper functions with isolation validation
-  const setCacheEntry = useCallback(
-    (
-      spawnId: string | undefined,
-      spawnAssetId: string | undefined,
-      draft: { draftValues: Partial<MediaAssetProperties> },
-    ) => {
-      validateCacheScope(spawnId, spawnAssetId, "set");
-      validateAssetScope(spawnAssetId, "set");
-
-      const key = generateCacheKey(spawnId, spawnAssetId);
-      if (process.env.NODE_ENV === "development") {
-        console.debug("Setting cache entry:", key, draft);
-      }
-      assetDraftCacheRef.current = {
-        ...assetDraftCacheRef.current,
-        [key]: draft,
-      };
-    },
-    [generateCacheKey, validateCacheScope, validateAssetScope],
-  );
-
-  const getCacheEntry = useCallback(
-    (spawnId: string | undefined, spawnAssetId: string | undefined) => {
-      validateCacheScope(spawnId, spawnAssetId, "get");
-      validateAssetScope(spawnAssetId, "get");
-
-      const key = generateCacheKey(spawnId, spawnAssetId);
-      const entry = assetDraftCacheRef.current[key];
-      if (process.env.NODE_ENV === "development") {
-        console.debug(
-          "Getting cache entry:",
-          key,
-          entry ? "found" : "not found",
-        );
-      }
-      return entry;
-    },
-    [generateCacheKey, validateCacheScope, validateAssetScope],
-  );
-
-  const clearCacheEntry = useCallback(
-    (spawnId: string | undefined, spawnAssetId: string | undefined) => {
-      validateCacheScope(spawnId, spawnAssetId, "clear");
-      validateAssetScope(spawnAssetId, "clear");
-
-      const key = generateCacheKey(spawnId, spawnAssetId);
-      if (process.env.NODE_ENV === "development") {
-        console.debug("Clearing asset draft cache entry:", key);
-      }
-      const newCache = { ...assetDraftCacheRef.current };
-      delete newCache[key];
-      assetDraftCacheRef.current = newCache;
-    },
-    [generateCacheKey, validateCacheScope, validateAssetScope],
-  );
-
-  // Special cache operations for system-level clearing (bypasses asset scope validation)
-  const clearCacheEntryUnsafe = useCallback(
-    (spawnId: string | undefined, spawnAssetId: string | undefined) => {
-      // Only validate cache scope, not asset scope (for system operations)
-      validateCacheScope(spawnId, spawnAssetId, "clear-unsafe");
-
-      const key = generateCacheKey(spawnId, spawnAssetId);
-      if (process.env.NODE_ENV === "development") {
-        console.debug(
-          "Clearing asset draft cache entry (unsafe system operation):",
-          key,
-        );
-      }
-      const newCache = { ...assetDraftCacheRef.current };
-      delete newCache[key];
-      assetDraftCacheRef.current = newCache;
-    },
-    [generateCacheKey, validateCacheScope],
-  );
-
-  const clearAllCacheEntries = useCallback(() => {
-    if (process.env.NODE_ENV === "development") {
-      console.debug("Clearing all asset draft cache entries");
-    }
-    assetDraftCacheRef.current = {};
-  }, []);
-
-  const [showMetadata, setShowMetadata] = useState<boolean>(true);
   const [isTesting, setIsTesting] = useState<boolean>(false);
   const validation = useMemo(() => validateTrigger(trigger), [trigger]);
 
@@ -828,14 +510,15 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
     }
   }, [selectedSpawn, isTesting]);
 
-  const formatDate = (ms: number | undefined) => {
-    if (!ms) return "-";
-    try {
-      return new Date(ms).toLocaleString();
-    } catch {
-      return String(ms);
-    }
-  };
+  const handleTriggerTypeChange = useCallback((nextType: TriggerType) => {
+    pendingTriggerTypeRef.current = nextType;
+    setShowTriggerTypeDialog(true);
+  }, []);
+
+  const handleDelete = useCallback(() => {
+    setDeleteError(null);
+    setShowDeleteDialog(true);
+  }, []);
 
   if (!selectedSpawnId) {
     return (
@@ -860,171 +543,45 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
   // Render all confirmation dialogs at the top level so they're always in the render tree
   return (
     <>
-      <ConfirmDialog
-        isOpen={showDiscardDialog}
-        title="Discard Unsaved Changes?"
-        message="You have unsaved changes. If you discard now, your edits will be lost."
-        confirmText="Discard changes"
-        cancelText="Keep editing"
-        variant="warning"
-        onConfirm={() => {
-          if (!selectedSpawn) {
-            setShowDiscardDialog(false);
-            return;
-          }
-
-          // Clear cache when discarding changes to prevent stale data
-          if (process.env.NODE_ENV === "development") {
-            console.debug(
-              "Discarding unsaved changes, clearing asset draft cache to prevent stale data",
-            );
-          }
-          clearAllCacheEntries();
-
-          setName(selectedSpawn.name);
-          setDescription(selectedSpawn.description || "");
-          setEnabled(!!selectedSpawn.enabled);
-          setDuration(selectedSpawn.duration);
-          setSaveError(null);
-          setSaveSuccess(null);
-          setShowDiscardDialog(false);
-        }}
-        onCancel={() => setShowDiscardDialog(false)}
-      />
-      <ConfirmDialog
-        isOpen={showTriggerTypeDialog}
-        title="Change Trigger Type?"
-        message="Changing the trigger type will reset the current trigger configuration."
-        confirmText="Change type"
-        cancelText="Cancel"
-        variant="warning"
-        onConfirm={() => {
-          const nextType = pendingTriggerTypeRef.current;
-          setShowTriggerTypeDialog(false);
-          if (!nextType) return;
-          setTrigger(getDefaultTrigger(nextType));
-          pendingTriggerTypeRef.current = null;
-        }}
-        onCancel={() => {
-          pendingTriggerTypeRef.current = null;
-          setShowTriggerTypeDialog(false);
-        }}
-      />
-      <ConfirmDialog
-        isOpen={showModeSwitchDialog}
-        title={getModalContent().title}
-        message={getModalContent().message}
-        confirmText="Switch"
-        cancelText="Stay"
-        variant="warning"
-        onConfirm={() => {
-          const pending = switchPendingRef.current;
-          setShowModeSwitchDialog(false);
-          if (!pending) return;
-
-          // Clear cache when switching modes to prevent contamination
-          if (process.env.NODE_ENV === "development") {
-            console.debug(
-              "Mode switch confirmed, clearing asset draft cache to prevent contamination between modes",
-            );
-          }
-          clearAllCacheEntries();
-
-          if (pending.mode === "asset-settings" && pending.spawnAssetId) {
-            selectSpawnAsset(pending.spawnAssetId);
-            setCenterPanelMode("asset-settings");
-          } else if (pending.mode === "spawn-settings") {
-            setCenterPanelMode("spawn-settings");
-            selectSpawnAsset(undefined);
-          }
-          switchPendingRef.current = null;
-        }}
-        onCancel={() => {
-          switchPendingRef.current = null;
-          setShowModeSwitchDialog(false);
-        }}
-      />
-      <ConfirmDialog
-        isOpen={showAssetSwitchDialog}
-        title="Unsaved Asset Settings"
-        message="Switching will discard property changes. Continue?"
-        confirmText="Switch"
-        cancelText="Stay"
-        variant="warning"
-        onConfirm={() => {
-          const pendingId = pendingAssetSwitchIdRef.current;
-          setShowAssetSwitchDialog(false);
-          setUnsavedChanges(false, "none");
-
-          // Clear the old asset's cached draft to prevent state bleed
-          if (selectedSpawnAssetId) {
-            clearCacheEntryUnsafe(selectedSpawnId, selectedSpawnAssetId);
-          }
-
-          // Clear the target asset's cached draft to prevent stale data
-          if (pendingId) {
-            clearCacheEntryUnsafe(selectedSpawnId, pendingId);
-          }
-
-          if (pendingId) {
-            selectSpawnAsset(pendingId);
-            prevSelectedSpawnAssetIdRef.current = pendingId;
-          }
-          pendingAssetSwitchIdRef.current = null;
-        }}
-        onCancel={() => {
-          pendingAssetSwitchIdRef.current = null;
-          setShowAssetSwitchDialog(false);
-        }}
-      />
-      <ConfirmDialog
-        isOpen={showDeleteDialog}
-        title="Delete Spawn?"
-        message="This action cannot be undone. The spawn will be permanently deleted."
-        confirmText="Delete"
-        cancelText="Cancel"
-        variant="danger"
-        onConfirm={async () => {
-          if (!selectedSpawn) {
-            setShowDeleteDialog(false);
-            return;
-          }
-          try {
-            const res = await SpawnService.deleteSpawn(selectedSpawn.id);
-            if (!res.success) {
-              setDeleteError(res.error || "Failed to delete spawn");
-              setShowDeleteDialog(false);
-              return;
-            }
-
-            // Clear cache when deleting spawn to prevent stale data
-            if (process.env.NODE_ENV === "development") {
-              console.debug(
-                "Spawn deleted, clearing asset draft cache to prevent stale data",
-              );
-            }
-            clearAllCacheEntries();
-
-            // Notify list to remove and clear selection
-            dispatchMediaSpawnerEvent(MediaSpawnerEvents.SPAWN_DELETED, {
-              id: selectedSpawn.id,
-            });
-            setSelectedSpawn(null);
-            setName("");
-            setDescription("");
-            setSaveError(null);
-            setSaveSuccess(null);
-            setShowDeleteDialog(false);
-            // Clear selection in layout context so center panel returns to guidance
-            selectSpawn(undefined);
-          } catch (e) {
-            setDeleteError(
-              e instanceof Error ? e.message : "Failed to delete spawn",
-            );
-            setShowDeleteDialog(false);
-          }
-        }}
-        onCancel={() => setShowDeleteDialog(false)}
+      <SpawnEditorDialogs
+        // Discard dialog
+        showDiscardDialog={showDiscardDialog}
+        setShowDiscardDialog={setShowDiscardDialog}
+        selectedSpawn={selectedSpawn}
+        clearAllCacheEntries={clearAllCacheEntries}
+        setName={setName}
+        setDescription={setDescription}
+        setEnabled={setEnabled}
+        setDuration={setDuration}
+        setSaveError={setSaveError}
+        setSaveSuccess={setSaveSuccess}
+        // Trigger type dialog
+        showTriggerTypeDialog={showTriggerTypeDialog}
+        setShowTriggerTypeDialog={setShowTriggerTypeDialog}
+        pendingTriggerTypeRef={pendingTriggerTypeRef}
+        setTrigger={setTrigger}
+        // Mode switch dialog
+        showModeSwitchDialog={showModeSwitchDialog}
+        setShowModeSwitchDialog={setShowModeSwitchDialog}
+        getModalContent={getModalContent}
+        switchPendingRef={switchPendingRef}
+        selectSpawnAsset={selectSpawnAsset}
+        setCenterPanelMode={setCenterPanelMode}
+        // Asset switch dialog
+        showAssetSwitchDialog={showAssetSwitchDialog}
+        setShowAssetSwitchDialog={setShowAssetSwitchDialog}
+        pendingAssetSwitchIdRef={pendingAssetSwitchIdRef}
+        setUnsavedChanges={setUnsavedChanges}
+        selectedSpawnAssetId={selectedSpawnAssetId}
+        selectedSpawnId={selectedSpawnId}
+        clearCacheEntryUnsafe={clearCacheEntryUnsafe}
+        prevSelectedSpawnAssetIdRef={prevSelectedSpawnAssetIdRef}
+        // Delete dialog
+        showDeleteDialog={showDeleteDialog}
+        setShowDeleteDialog={setShowDeleteDialog}
+        setDeleteError={setDeleteError}
+        setSelectedSpawn={setSelectedSpawn}
+        selectSpawn={selectSpawn}
       />
 
       {/* Conditionally render Asset Settings or Spawn Editor */}
@@ -1056,75 +613,18 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
         />
       ) : (
         <div className="h-full flex flex-col">
-          <div className="p-4 border-b border-[rgb(var(--color-border))] bg-[rgb(var(--color-muted))]/5">
-            <div className="flex items-center gap-2">
-              <h2 className="text-lg font-semibold text-[rgb(var(--color-fg))]">
-                Spawn Editor
-              </h2>
-              {hasUnsavedChanges && (
-                <span className="text-[rgb(var(--color-warning))] text-sm font-medium">
-                  • Unsaved changes
-                </span>
-              )}
-            </div>
-            <div className="flex flex-col md:flex-row md:items-center md:justify-between gap-3 mt-1">
-              <p className="text-sm text-[rgb(var(--color-muted-foreground))]">
-                {selectedSpawn
-                  ? `Editing: ${selectedSpawn.name}`
-                  : "Loading spawn..."}
-              </p>
-              <div className="flex items-center gap-2">
-                {selectedSpawn && (
-                  <>
-                    <Button
-                      type="button"
-                      onClick={handleTestSpawn}
-                      variant="outline"
-                      size="sm"
-                      disabled={isTesting}
-                      loading={isTesting}
-                      aria-label="Test spawn"
-                    >
-                      Test
-                    </Button>
-                    <Button
-                      type="button"
-                      onClick={() => {
-                        setDeleteError(null);
-                        setShowDeleteDialog(true);
-                      }}
-                      variant="destructive"
-                      size="sm"
-                      aria-label="Delete spawn"
-                    >
-                      Delete
-                    </Button>
-                  </>
-                )}
-                <Button
-                  type="button"
-                  onClick={handleCancel}
-                  variant="outline"
-                  size="sm"
-                  disabled={isCancelDisabled}
-                  aria-label="Cancel edits"
-                >
-                  Cancel
-                </Button>
-                <Button
-                  type="button"
-                  onClick={handleSave}
-                  disabled={isSaveDisabled}
-                  variant="primary"
-                  size="sm"
-                  loading={isSaving}
-                  aria-label="Save spawn"
-                >
-                  {isSaving ? "Saving..." : "Save"}
-                </Button>
-              </div>
-            </div>
-          </div>
+          <SpawnSettingsHeader
+            selectedSpawn={selectedSpawn}
+            hasUnsavedChanges={hasUnsavedChanges}
+            isTesting={isTesting}
+            isSaving={isSaving}
+            isSaveDisabled={isSaveDisabled}
+            isCancelDisabled={isCancelDisabled}
+            onTest={handleTestSpawn}
+            onDelete={handleDelete}
+            onCancel={handleCancel}
+            onSave={handleSave}
+          />
           <div className="flex-1 p-4">
             {deleteError && (
               <div className="mb-4 p-3 bg-[rgb(var(--color-error-bg))] border border-[rgb(var(--color-error-border))] text-sm text-[rgb(var(--color-error))] rounded">
@@ -1143,812 +643,61 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
             )}
             {selectedSpawn ? (
               <div className="max-w-2xl space-y-5">
-                <Card>
-                  <CardHeader>
-                    <div className="flex items-center justify-between">
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))]">
-                        Basic Details
-                      </CardTitle>
-                      <div className="flex items-center gap-2">
-                        <Switch
-                          id="spawn-enabled"
-                          checked={enabled}
-                          onCheckedChange={setEnabled}
-                          aria-label="Enabled"
-                        />
-                        <label
-                          htmlFor="spawn-enabled"
-                          className="text-sm text-[rgb(var(--color-fg))] cursor-pointer"
-                        >
-                          Enabled
-                        </label>
-                      </div>
-                    </div>
-                  </CardHeader>
-                  <CardContent>
-                    <div className="mb-2 flex items-center gap-2 text-xs">
-                      {validation.errors.length > 0 ? (
-                        <span className="px-2 py-0.5 rounded bg-[rgb(var(--color-error-bg))] text-[rgb(var(--color-error))]">
-                          Invalid
-                        </span>
-                      ) : validation.warnings.length > 0 ? (
-                        <span className="px-2 py-0.5 rounded bg-[rgb(var(--color-warning))]/10 text-[rgb(var(--color-warning))]">
-                          Warning
-                        </span>
-                      ) : (
-                        <span className="px-2 py-0.5 rounded bg-[rgb(var(--color-success))]/10 text-[rgb(var(--color-success))]">
-                          Valid
-                        </span>
-                      )}
-                      {validation.warnings.length > 0 && (
-                        <span className="text-[rgb(var(--color-muted-foreground))]">
-                          {validation.warnings[0]}
-                        </span>
-                      )}
-                    </div>
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                      <div>
-                        <Input
-                          id="spawn-name"
-                          type="text"
-                          label="Name"
-                          value={name}
-                          onChange={(e) => setName(e.target.value)}
-                          error={
-                            !trimmedName
-                              ? "Name is required"
-                              : !isNameValid
-                                ? "Name must be unique"
-                                : undefined
-                          }
-                        />
-                      </div>
+                <BasicDetailsCard
+                  name={name}
+                  setName={setName}
+                  description={description}
+                  setDescription={setDescription}
+                  duration={duration}
+                  setDuration={setDuration}
+                  enabled={enabled}
+                  setEnabled={setEnabled}
+                  trimmedName={trimmedName}
+                  isNameValid={isNameValid}
+                  validation={validation}
+                />
 
-                      <div>
-                        <Input
-                          id="spawn-duration"
-                          type="number"
-                          label="Duration (ms)"
-                          value={duration}
-                          onChange={(e) => {
-                            const val = Number(e.target.value);
-                            setDuration(val >= 0 ? val : 0);
-                          }}
-                          min={0}
-                          error={
-                            duration < 0
-                              ? "Duration must be non-negative"
-                              : undefined
-                          }
-                        />
-                      </div>
-
-                      <div className="md:col-span-2">
-                        <Input
-                          id="spawn-description"
-                          type="text"
-                          label="Description"
-                          value={description}
-                          onChange={(e) => setDescription(e.target.value)}
-                        />
-                      </div>
-                    </div>
-                  </CardContent>
-                </Card>
-
-                <Card>
-                  <CardHeader>
-                    <div className="flex items-center justify-between">
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))]">
-                        Trigger
-                      </CardTitle>
-                      <div className="flex items-center gap-2">
-                        <Switch
-                          checked={trigger?.enabled !== false}
-                          onCheckedChange={(checked) => {
-                            if (!trigger) return;
-                            setTrigger({ ...trigger, enabled: checked });
-                          }}
-                          aria-label="Trigger Enabled"
-                        />
-                        <label className="text-sm text-[rgb(var(--color-fg))] cursor-pointer">
-                          Trigger Enabled
-                        </label>
-                      </div>
-                    </div>
-                  </CardHeader>
-                  <CardContent>
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                      <div>
-                        <Input
-                          id="trigger-type"
-                          type="select"
-                          label="Trigger Type"
-                          value={trigger?.type || "manual"}
-                          onChange={(e) => {
-                            const nextType = e.target.value as TriggerType;
-                            if (!trigger) {
-                              setTrigger(getDefaultTrigger(nextType));
-                              return;
-                            }
-                            if (nextType === trigger.type) return;
-                            pendingTriggerTypeRef.current = nextType;
-                            setShowTriggerTypeDialog(true);
-                          }}
-                        >
-                          <option value="manual">Manual</option>
-                          <option value="time.atDateTime">
-                            Time: At Date/Time
-                          </option>
-                          <option value="time.dailyAt">Time: Daily At</option>
-                          <option value="time.everyNMinutes">
-                            Time: Every N Minutes
-                          </option>
-                          <option value="time.weeklyAt">Time: Weekly At</option>
-                          <option value="time.monthlyOn">
-                            Time: Monthly On
-                          </option>
-                          <option value="time.minuteOfHour">
-                            Time: Minute Of Hour
-                          </option>
-                          <option value="streamerbot.command">
-                            Streamer.bot Command
-                          </option>
-                          <option value="twitch.channelPointReward">
-                            Twitch: Channel Point Reward
-                          </option>
-                          <option value="twitch.subscription">
-                            Twitch: Subscription
-                          </option>
-                          <option value="twitch.giftSub">
-                            Twitch: Gifted Subs
-                          </option>
-                          <option value="twitch.cheer">Twitch: Cheer</option>
-                          <option value="twitch.follow">Twitch: Follow</option>
-                        </Input>
-                      </div>
-                      <div className="md:col-span-2 text-xs text-[rgb(var(--color-muted-foreground))]">
-                        {(() => {
-                          const t = trigger?.type || "manual";
-                          if (t === "manual") {
-                            return (
-                              <p>
-                                Manual triggers are activated outside of
-                                MediaSpawner. No configuration required.
-                              </p>
-                            );
-                          }
-                          if (
-                            t === "time.atDateTime" ||
-                            t === "time.dailyAt" ||
-                            t === "time.everyNMinutes" ||
-                            t === "time.minuteOfHour" ||
-                            t === "time.weeklyAt" ||
-                            t === "time.monthlyOn"
-                          ) {
-                            return (
-                              <p>
-                                Schedule triggers based on timezone-aware
-                                date/time rules.
-                              </p>
-                            );
-                          }
-                          if (t === "streamerbot.command") {
-                            return (
-                              <p>
-                                Streamer.bot command: configure command aliases
-                                and platform sources.
-                              </p>
-                            );
-                          }
-                          if (t === "twitch.channelPointReward") {
-                            return (
-                              <p>
-                                Twitch Channel Point reward redemption.
-                                Configure reward details in a later story.
-                              </p>
-                            );
-                          }
-                          if (t === "twitch.subscription") {
-                            return (
-                              <p>
-                                Configure subscription tier or minimum months.
-                              </p>
-                            );
-                          }
-                          if (t === "twitch.giftSub") {
-                            return (
-                              <p>
-                                Configure gifted sub count and optional tier.
-                              </p>
-                            );
-                          }
-                          if (t === "twitch.cheer") {
-                            return (
-                              <p>Configure minimum bits required to trigger.</p>
-                            );
-                          }
-                          if (t === "twitch.follow") {
-                            return (
-                              <p>
-                                Twitch follow events. No additional
-                                configuration required.
-                              </p>
-                            );
-                          }
-                          return null;
-                        })()}
-                      </div>
-                    </div>
-                  </CardContent>
-                </Card>
+                <TriggerTypeSelector
+                  trigger={trigger}
+                  setTrigger={setTrigger}
+                  onTriggerTypeChange={handleTriggerTypeChange}
+                />
 
                 {trigger?.type === "streamerbot.command" && (
-                  <Card>
-                    <CardHeader>
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))] mb-3">
-                        Command Configuration
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="space-y-4">
-                        {/* Command Aliases */}
-                        <div>
-                          <label className="block text-sm font-medium text-[rgb(var(--color-fg))] mb-2">
-                            Command Aliases
-                            {/* Refresh moved into SBCommandAliasCombobox */}
-                          </label>
-                          <div className="space-y-2">
-                            {(getCommandConfig(trigger)?.aliases || [""]).map(
-                              (alias: string, index: number) => (
-                                <div
-                                  key={index}
-                                  className="flex items-center gap-2"
-                                >
-                                  <SBCommandAliasCombobox
-                                    value={alias}
-                                    onChange={(v) => {
-                                      const newAliases = [
-                                        ...(getCommandConfig(trigger)
-                                          ?.aliases || [""]),
-                                      ];
-                                      newAliases[index] =
-                                        typeof v === "string" ? v : "";
-                                      setTrigger({
-                                        ...trigger,
-                                        config: {
-                                          ...getCommandConfig(trigger),
-                                          aliases: newAliases,
-                                        },
-                                      });
-                                    }}
-                                  />
-                                  {(getCommandConfig(trigger)?.aliases || [""])
-                                    .length > 1 && (
-                                    <button
-                                      type="button"
-                                      onClick={() => {
-                                        const newAliases = (
-                                          getCommandConfig(trigger)
-                                            ?.aliases || [""]
-                                        ).filter(
-                                          (_: string, i: number) => i !== index,
-                                        );
-                                        setTrigger({
-                                          ...trigger,
-                                          config: {
-                                            ...getCommandConfig(trigger),
-                                            aliases: newAliases,
-                                          },
-                                        });
-                                      }}
-                                      className="px-2 py-1 text-[rgb(var(--color-error))] hover:text-[rgb(var(--color-error-hover))] hover:bg-[rgb(var(--color-error-bg))] rounded"
-                                      aria-label="Remove command alias"
-                                    >
-                                      Remove
-                                    </button>
-                                  )}
-                                </div>
-                              ),
-                            )}
-                            <button
-                              type="button"
-                              onClick={() => {
-                                const newAliases = [
-                                  ...(getCommandConfig(trigger)?.aliases || [
-                                    "",
-                                  ]),
-                                  "",
-                                ];
-                                setTrigger({
-                                  ...trigger,
-                                  config: {
-                                    ...getCommandConfig(trigger),
-                                    aliases: newAliases,
-                                  },
-                                });
-                              }}
-                              className="px-3 py-1 text-sm text-[rgb(var(--color-accent))] hover:text-[rgb(var(--color-accent-hover))] hover:bg-[rgb(var(--color-accent))]/10 rounded border border-[rgb(var(--color-accent))]/20"
-                            >
-                              + Add Alias
-                            </button>
-                          </div>
-                          {(() => {
-                            const config = getCommandConfig(trigger);
-                            const aliases = config?.aliases || [];
-                            const hasEmptyAlias = aliases.some(
-                              (a: string) => !a.trim(),
-                            );
-                            if (!aliases.length || hasEmptyAlias) {
-                              return (
-                                <p className="mt-1 text-xs text-[rgb(var(--color-error))]">
-                                  At least one command alias is required
-                                </p>
-                              );
-                            }
-                            return null;
-                          })()}
-                        </div>
-
-                        {/* Case Sensitivity */}
-                        <div>
-                          <div className="flex items-center gap-2">
-                            <Switch
-                              aria-label="Case sensitive"
-                              checked={
-                                getCommandConfig(trigger)?.caseSensitive ||
-                                false
-                              }
-                              onCheckedChange={(checked) => {
-                                setTrigger({
-                                  ...trigger,
-                                  config: {
-                                    ...getCommandConfig(trigger),
-                                    caseSensitive: checked,
-                                  },
-                                });
-                              }}
-                            />
-                            <label className="text-sm text-[rgb(var(--color-fg))] cursor-pointer">
-                              Case sensitive
-                            </label>
-                          </div>
-                          <p className="mt-1 text-xs text-[rgb(var(--color-muted-foreground))]">
-                            When enabled, command matching will be
-                            case-sensitive
-                          </p>
-                        </div>
-
-                        {/* Platform Sources removed: Twitch-only support for now */}
-
-                        {/* Filtering Options */}
-                        <div className="space-y-3">
-                          <div>
-                            <div className="flex items-center gap-2">
-                              <Switch
-                                aria-label="Ignore internal messages"
-                                checked={
-                                  getCommandConfig(trigger)?.ignoreInternal !==
-                                  false
-                                }
-                                onCheckedChange={(checked) => {
-                                  setTrigger({
-                                    ...trigger,
-                                    config: {
-                                      ...getCommandConfig(trigger),
-                                      ignoreInternal: checked,
-                                    },
-                                  });
-                                }}
-                              />
-                              <label className="text-sm text-[rgb(var(--color-fg))] cursor-pointer">
-                                Ignore internal messages
-                              </label>
-                            </div>
-                            <p className="mt-1 text-xs text-[rgb(var(--color-muted-foreground))]">
-                              Skip messages from internal/system sources
-                            </p>
-                          </div>
-
-                          <div>
-                            <div className="flex items-center gap-2">
-                              <Switch
-                                aria-label="Ignore bot account messages"
-                                checked={
-                                  getCommandConfig(trigger)
-                                    ?.ignoreBotAccount !== false
-                                }
-                                onCheckedChange={(checked) => {
-                                  setTrigger({
-                                    ...trigger,
-                                    config: {
-                                      ...getCommandConfig(trigger),
-                                      ignoreBotAccount: checked,
-                                    },
-                                  });
-                                }}
-                              />
-                              <label className="text-sm text-[rgb(var(--color-fg))] cursor-pointer">
-                                Ignore bot account messages
-                              </label>
-                            </div>
-                            <p className="mt-1 text-xs text-[rgb(var(--color-muted-foreground))]">
-                              Skip messages from the bot account to avoid loops
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
+                  <CommandTriggerConfig
+                    trigger={trigger}
+                    setTrigger={setTrigger}
+                  />
                 )}
 
                 {trigger?.type === "twitch.channelPointReward" && (
-                  <Card>
-                    <CardHeader>
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))] mb-3">
-                        Channel Point Reward Configuration
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="space-y-4">
-                        {/* Reward Identifier */}
-                        <div>
-                          <Input
-                            type="text"
-                            label="Reward Identifier"
-                            value={
-                              getChannelPointConfig(trigger)
-                                ?.rewardIdentifier || ""
-                            }
-                            onChange={(e) => {
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getChannelPointConfig(trigger),
-                                  rewardIdentifier: e.target.value,
-                                },
-                              });
-                            }}
-                            placeholder="Enter reward name or ID (e.g., Alert, Scene1, 12345)"
-                            error={
-                              !getChannelPointConfig(
-                                trigger,
-                              )?.rewardIdentifier?.trim()
-                                ? "Reward identifier is required"
-                                : undefined
-                            }
-                            helperText="Enter the name or ID of the channel point reward from your Twitch channel"
-                          />
-                        </div>
-
-                        {/* Use Viewer Input */}
-                        <div>
-                          <div className="flex items-center gap-2">
-                            <Switch
-                              aria-label="Use viewer input in spawn configuration"
-                              checked={
-                                getChannelPointConfig(trigger)
-                                  ?.useViewerInput || false
-                              }
-                              onCheckedChange={(checked) => {
-                                setTrigger({
-                                  ...trigger,
-                                  config: {
-                                    ...getChannelPointConfig(trigger),
-                                    useViewerInput: checked,
-                                  },
-                                });
-                              }}
-                            />
-                            <label className="text-sm text-[rgb(var(--color-fg))] cursor-pointer">
-                              Use viewer input in spawn configuration
-                            </label>
-                          </div>
-                          <p className="mt-1 text-xs text-[rgb(var(--color-muted-foreground))]">
-                            When enabled, the viewer's message will be available
-                            for use in spawn settings
-                          </p>
-                        </div>
-
-                        {/* Redemption Statuses */}
-                        <div>
-                          <label className="block text-sm font-medium text-[rgb(var(--color-fg))] mb-2">
-                            Redemption Statuses
-                          </label>
-                          <div className="space-y-2">
-                            {["pending", "fulfilled", "cancelled"].map(
-                              (status) => (
-                                <div
-                                  key={status}
-                                  className="flex items-center gap-2"
-                                >
-                                  <Switch
-                                    aria-label={status}
-                                    checked={(
-                                      getChannelPointConfig(trigger)
-                                        ?.statuses || ["fulfilled"]
-                                    ).includes(status)}
-                                    onCheckedChange={(checked) => {
-                                      const currentStatuses =
-                                        getChannelPointConfig(trigger)
-                                          ?.statuses || ["fulfilled"];
-                                      const newStatuses = checked
-                                        ? [...currentStatuses, status]
-                                        : currentStatuses.filter(
-                                            (s: string) => s !== status,
-                                          );
-                                      setTrigger({
-                                        ...trigger,
-                                        config: {
-                                          ...getChannelPointConfig(trigger),
-                                          statuses: newStatuses,
-                                        },
-                                      });
-                                    }}
-                                  />
-                                  <label className="text-sm text-[rgb(var(--color-fg))] cursor-pointer capitalize">
-                                    {status}
-                                  </label>
-                                </div>
-                              ),
-                            )}
-                          </div>
-                          {(() => {
-                            const config = getChannelPointConfig(trigger);
-                            const statuses = config?.statuses || [];
-                            if (statuses.length === 0) {
-                              return (
-                                <p className="mt-1 text-xs text-[rgb(var(--color-error))]">
-                                  At least one redemption status must be
-                                  selected
-                                </p>
-                              );
-                            }
-                            return null;
-                          })()}
-                          <p className="mt-1 text-xs text-[rgb(var(--color-muted-foreground))]">
-                            Select which redemption statuses should trigger this
-                            spawn
-                          </p>
-                        </div>
-
-                        {/* Help Text */}
-                        <div className="bg-[rgb(var(--color-accent))]/10 border border-[rgb(var(--color-accent))]/20 rounded-lg p-3">
-                          <p className="text-xs text-[rgb(var(--color-accent-text))]">
-                            <strong>Note:</strong> Twitch handles all reward
-                            logic including cooldowns, usage limits, and point
-                            costs. MediaSpawner only configures when spawns
-                            trigger based on redemption events.
-                          </p>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
+                  <ChannelPointTriggerConfig
+                    trigger={trigger}
+                    setTrigger={setTrigger}
+                  />
                 )}
 
                 {trigger?.type === "twitch.subscription" && (
-                  <Card>
-                    <CardHeader>
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))] mb-3">
-                        Subscription Configuration
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                        <div>
-                          <Input
-                            id="sub-tier"
-                            type="select"
-                            label="Tier"
-                            value={getSubscriptionConfig(trigger)?.tier ?? ""}
-                            onChange={(e) => {
-                              const v = (e.target.value || undefined) as
-                                | "1000"
-                                | "2000"
-                                | "3000"
-                                | undefined;
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getSubscriptionConfig(trigger),
-                                  tier: v,
-                                },
-                              });
-                            }}
-                          >
-                            <option value="">Any</option>
-                            <option value="1000">Tier 1</option>
-                            <option value="2000">Tier 2</option>
-                            <option value="3000">Tier 3</option>
-                          </Input>
-                        </div>
-                        <div>
-                          <Input
-                            id="sub-months-comparator"
-                            type="select"
-                            label="Months Comparator"
-                            value={
-                              getSubscriptionConfig(trigger)
-                                ?.monthsComparator ?? ""
-                            }
-                            onChange={(e) => {
-                              const v = (e.target.value || undefined) as
-                                | "lt"
-                                | "eq"
-                                | "gt"
-                                | undefined;
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getSubscriptionConfig(trigger),
-                                  monthsComparator: v,
-                                },
-                              });
-                            }}
-                            error={validation.fieldErrors.monthsComparator?.[0]}
-                          >
-                            <option value="">- Select -</option>
-                            <option value="lt">Less than</option>
-                            <option value="eq">Equal to</option>
-                            <option value="gt">Greater than</option>
-                          </Input>
-                        </div>
-                        <div>
-                          <Input
-                            id="sub-months"
-                            type="number"
-                            label="Months"
-                            min={1}
-                            value={getSubscriptionConfig(trigger)?.months ?? ""}
-                            onChange={(e) => {
-                              const val = Math.max(
-                                1,
-                                Number(e.target.value) || 1,
-                              );
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getSubscriptionConfig(trigger),
-                                  months: val,
-                                },
-                              });
-                            }}
-                            error={validation.fieldErrors.months?.[0]}
-                          />
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
+                  <SubscriptionTriggerConfig
+                    trigger={trigger}
+                    setTrigger={setTrigger}
+                    validation={validation}
+                  />
                 )}
 
                 {trigger?.type === "twitch.giftSub" && (
-                  <Card>
-                    <CardHeader>
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))] mb-3">
-                        Gifted Subs Configuration
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                        <div>
-                          <Input
-                            id="gift-min-count"
-                            type="number"
-                            label="Minimum Count"
-                            min={1}
-                            value={getGiftSubConfig(trigger)?.minCount ?? ""}
-                            onChange={(e) => {
-                              const val = Math.max(
-                                1,
-                                Number(e.target.value) || 1,
-                              );
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getGiftSubConfig(trigger),
-                                  minCount: val,
-                                },
-                              });
-                            }}
-                          />
-                        </div>
-                        <div>
-                          <Input
-                            id="gift-tier"
-                            type="select"
-                            label="Tier"
-                            value={getGiftSubConfig(trigger)?.tier ?? ""}
-                            onChange={(e) => {
-                              const v = (e.target.value || undefined) as
-                                | "1000"
-                                | "2000"
-                                | "3000"
-                                | undefined;
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getGiftSubConfig(trigger),
-                                  tier: v,
-                                },
-                              });
-                            }}
-                          >
-                            <option value="">Any</option>
-                            <option value="1000">Tier 1</option>
-                            <option value="2000">Tier 2</option>
-                            <option value="3000">Tier 3</option>
-                          </Input>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
+                  <GiftSubTriggerConfig
+                    trigger={trigger}
+                    setTrigger={setTrigger}
+                  />
                 )}
 
                 {trigger?.type === "twitch.cheer" && (
-                  <Card>
-                    <CardHeader>
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))] mb-3">
-                        Cheer Configuration
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                        <div>
-                          <Input
-                            id="cheer-bits-comparator"
-                            type="select"
-                            label="Bits Comparator"
-                            value={
-                              getCheerConfig(trigger)?.bitsComparator ?? ""
-                            }
-                            onChange={(e) => {
-                              const v = (e.target.value || undefined) as
-                                | "lt"
-                                | "eq"
-                                | "gt"
-                                | undefined;
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getCheerConfig(trigger),
-                                  bitsComparator: v,
-                                },
-                              });
-                            }}
-                            error={validation.fieldErrors.bitsComparator?.[0]}
-                          >
-                            <option value="">- Select -</option>
-                            <option value="lt">Less than</option>
-                            <option value="eq">Equal to</option>
-                            <option value="gt">Greater than</option>
-                          </Input>
-                        </div>
-                        <div>
-                          <Input
-                            id="cheer-bits"
-                            type="number"
-                            label="Bits"
-                            min={1}
-                            value={getCheerConfig(trigger)?.bits ?? ""}
-                            onChange={(e) => {
-                              const val = Math.max(
-                                1,
-                                Number(e.target.value) || 1,
-                              );
-                              setTrigger({
-                                ...trigger,
-                                config: {
-                                  ...getCheerConfig(trigger),
-                                  bits: val,
-                                },
-                              });
-                            }}
-                            error={validation.fieldErrors.bits?.[0]}
-                          />
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
+                  <CheerTriggerConfig
+                    trigger={trigger}
+                    setTrigger={setTrigger}
+                    validation={validation}
+                  />
                 )}
 
                 {(trigger?.type === "time.atDateTime" ||
@@ -1957,793 +706,11 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
                   trigger?.type === "time.minuteOfHour" ||
                   trigger?.type === "time.weeklyAt" ||
                   trigger?.type === "time.monthlyOn") && (
-                  <Card>
-                    <CardHeader>
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))] mb-3">
-                        Time-based Configuration
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="space-y-4">
-                        {(() => {
-                          const next = getNextActivation(trigger);
-                          return (
-                            <div className="bg-[rgb(var(--color-muted))]/5 border border-[rgb(var(--color-border))] rounded p-2 text-sm text-[rgb(var(--color-fg))]">
-                              <span className="font-medium">
-                                Next activation:{" "}
-                              </span>
-                              {formatNextActivation(next.when, next.timezone)}
-                            </div>
-                          );
-                        })()}
-                        {trigger?.type === "time.weeklyAt" && (
-                          <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                            <div className="md:col-span-1">
-                              <fieldset className="space-y-2">
-                                <legend className="text-sm font-medium text-[rgb(var(--color-fg))] mb-2">
-                                  Days of Week
-                                </legend>
-                                <div className="grid grid-cols-2 md:grid-cols-4 lg:grid-cols-7 gap-2">
-                                  {dayOfWeekOptions.map((d) => {
-                                    const config = getWeeklyAtConfig(trigger);
-                                    const daysOfWeek = config?.daysOfWeek ?? [
-                                      1,
-                                    ];
-                                    const isChecked = daysOfWeek.includes(
-                                      d.value,
-                                    );
-                                    const isLastDay =
-                                      daysOfWeek.length === 1 && isChecked;
-                                    return (
-                                      <label
-                                        key={d.value}
-                                        className="flex items-center space-x-2 cursor-pointer"
-                                      >
-                                        <input
-                                          type="checkbox"
-                                          checked={isChecked}
-                                          disabled={
-                                            trigger?.enabled === false ||
-                                            isLastDay
-                                          }
-                                          onChange={(e) => {
-                                            const base =
-                                              getWeeklyAtConfig(trigger) ||
-                                              ({
-                                                daysOfWeek: [1],
-                                                time: "09:00",
-                                                timezone: moment.tz.guess(),
-                                              } as {
-                                                daysOfWeek: number[];
-                                                time: string;
-                                                timezone: string;
-                                              });
-                                            const currentDays = base.daysOfWeek;
-                                            let newDays: number[];
-                                            if (e.target.checked) {
-                                              newDays = [
-                                                ...currentDays,
-                                                d.value,
-                                              ];
-                                            } else {
-                                              newDays = currentDays.filter(
-                                                (day) => day !== d.value,
-                                              );
-                                            }
-                                            setTrigger({
-                                              ...trigger!,
-                                              config: {
-                                                ...base,
-                                                daysOfWeek: newDays,
-                                              },
-                                            });
-                                          }}
-                                          className="rounded border-[rgb(var(--color-border))] text-[rgb(var(--color-accent))] focus-visible:ring-2 focus-visible:ring-[rgb(var(--color-ring))] focus-visible:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed"
-                                        />
-                                        <span className="text-sm text-[rgb(var(--color-fg))]">
-                                          {d.label}
-                                        </span>
-                                      </label>
-                                    );
-                                  })}
-                                </div>
-                              </fieldset>
-                            </div>
-                            <div>
-                              <Input
-                                type="time"
-                                label="Time (HH:mm)"
-                                value={
-                                  getWeeklyAtConfig(trigger)?.time || "09:00"
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getWeeklyAtConfig(trigger) ||
-                                    ({
-                                      daysOfWeek: [1],
-                                      time: "09:00",
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      daysOfWeek: number[];
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: { ...base, time: e.target.value },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                              />
-                            </div>
-                            <div>
-                              <Input
-                                type="select"
-                                label="Timezone"
-                                value={
-                                  getWeeklyAtConfig(trigger)?.timezone ||
-                                  moment.tz.guess()
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getWeeklyAtConfig(trigger) ||
-                                    ({
-                                      daysOfWeek: [1],
-                                      time: "09:00",
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      daysOfWeek: number[];
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: {
-                                      ...base,
-                                      timezone: e.target.value,
-                                    },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                              >
-                                {timezoneOptions.map((tz) => (
-                                  <option key={tz.value} value={tz.value}>
-                                    {tz.label}
-                                  </option>
-                                ))}
-                              </Input>
-                            </div>
-                          </div>
-                        )}
-                        {trigger?.type === "time.monthlyOn" && (
-                          <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                            <div>
-                              <Input
-                                type="number"
-                                label="Day of Month"
-                                min={1}
-                                max={31}
-                                value={
-                                  getMonthlyOnConfig(trigger)?.dayOfMonth ?? 1
-                                }
-                                onChange={(e) => {
-                                  const val = Math.max(
-                                    1,
-                                    Math.min(31, Number(e.target.value) || 1),
-                                  );
-                                  const base =
-                                    getMonthlyOnConfig(trigger) ||
-                                    ({
-                                      dayOfMonth: 1,
-                                      time: "09:00",
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      dayOfMonth: number;
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: { ...base, dayOfMonth: val },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                              />
-                            </div>
-                            <div>
-                              <Input
-                                type="time"
-                                label="Time (HH:mm)"
-                                value={
-                                  getMonthlyOnConfig(trigger)?.time || "09:00"
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getMonthlyOnConfig(trigger) ||
-                                    ({
-                                      dayOfMonth: 1,
-                                      time: "09:00",
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      dayOfMonth: number;
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: { ...base, time: e.target.value },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                              />
-                            </div>
-                            <div>
-                              <Input
-                                type="select"
-                                label="Timezone"
-                                value={
-                                  getMonthlyOnConfig(trigger)?.timezone ||
-                                  moment.tz.guess()
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getMonthlyOnConfig(trigger) ||
-                                    ({
-                                      dayOfMonth: 1,
-                                      time: "09:00",
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      dayOfMonth: number;
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: {
-                                      ...base,
-                                      timezone: e.target.value,
-                                    },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                              >
-                                {timezoneOptions.map((tz) => (
-                                  <option key={tz.value} value={tz.value}>
-                                    {tz.label}
-                                  </option>
-                                ))}
-                              </Input>
-                            </div>
-                          </div>
-                        )}
-                        {trigger?.type === "time.atDateTime" && (
-                          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                              <Input
-                                id="at-datetime"
-                                type="datetime-local"
-                                label="ISO Date-Time"
-                                value={(() => {
-                                  const cfg = getAtDateTimeConfig(trigger);
-                                  if (!cfg?.isoDateTime) return "";
-                                  try {
-                                    return moment(cfg.isoDateTime)
-                                      .tz(cfg.timezone)
-                                      .format("YYYY-MM-DDTHH:mm");
-                                  } catch {
-                                    return "";
-                                  }
-                                })()}
-                                onChange={(e) => {
-                                  const current = getAtDateTimeConfig(trigger);
-                                  const tz =
-                                    current?.timezone || moment.tz.guess();
-                                  const iso = e.target.value
-                                    ? moment
-                                        .tz(
-                                          e.target.value,
-                                          "YYYY-MM-DDTHH:mm",
-                                          tz,
-                                        )
-                                        .toISOString()
-                                    : new Date().toISOString();
-                                  const base =
-                                    current ||
-                                    ({
-                                      isoDateTime: new Date().toISOString(),
-                                      timezone: tz,
-                                    } as {
-                                      isoDateTime: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger,
-                                    config: { ...base, isoDateTime: iso },
-                                  });
-                                }}
-                                error={validation.fieldErrors.isoDateTime?.[0]}
-                              />
-                            </div>
-                            <div>
-                              <Input
-                                id="at-timezone"
-                                type="select"
-                                label="Timezone"
-                                value={
-                                  getAtDateTimeConfig(trigger)?.timezone ||
-                                  moment.tz.guess()
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getAtDateTimeConfig(trigger) ||
-                                    ({
-                                      isoDateTime: new Date().toISOString(),
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      isoDateTime: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: {
-                                      ...base,
-                                      timezone: e.target.value,
-                                    },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                                error={validation.fieldErrors.timezone?.[0]}
-                              >
-                                {timezoneOptions.map((tz) => (
-                                  <option key={tz.value} value={tz.value}>
-                                    {tz.label}
-                                  </option>
-                                ))}
-                              </Input>
-                            </div>
-                          </div>
-                        )}
-
-                        {trigger?.type === "time.dailyAt" && (
-                          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                              <Input
-                                id="daily-time"
-                                type="time"
-                                label="Time (HH:mm)"
-                                value={
-                                  getDailyAtConfig(trigger)?.time || "09:00"
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getDailyAtConfig(trigger) ||
-                                    ({ time: "09:00", timezone: "UTC" } as {
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger,
-                                    config: { ...base, time: e.target.value },
-                                  });
-                                }}
-                                error={validation.fieldErrors.time?.[0]}
-                              />
-                            </div>
-                            <div>
-                              <Input
-                                id="daily-timezone"
-                                type="select"
-                                label="Timezone"
-                                value={
-                                  getDailyAtConfig(trigger)?.timezone ||
-                                  moment.tz.guess()
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getDailyAtConfig(trigger) ||
-                                    ({
-                                      time: "09:00",
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      time: string;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: {
-                                      ...base,
-                                      timezone: e.target.value,
-                                    },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                                error={validation.fieldErrors.timezone?.[0]}
-                              >
-                                {timezoneOptions.map((tz) => (
-                                  <option key={tz.value} value={tz.value}>
-                                    {tz.label}
-                                  </option>
-                                ))}
-                              </Input>
-                            </div>
-                          </div>
-                        )}
-
-                        {trigger?.type === "time.everyNMinutes" && (
-                          <div className="space-y-4">
-                            <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                              <div>
-                                <Input
-                                  id="every-interval"
-                                  type="number"
-                                  label="Interval (minutes)"
-                                  min={1}
-                                  value={
-                                    getEveryNMinutesConfig(trigger)
-                                      ?.intervalMinutes ?? 15
-                                  }
-                                  onChange={(e) => {
-                                    const val = Math.max(
-                                      1,
-                                      Number(e.target.value) || 1,
-                                    );
-                                    const base =
-                                      getEveryNMinutesConfig(trigger) ||
-                                      ({
-                                        intervalMinutes: 15,
-                                        timezone: "UTC",
-                                      } as {
-                                        intervalMinutes: number;
-                                        timezone: string;
-                                        anchor?:
-                                          | { kind: "topOfHour" }
-                                          | {
-                                              kind: "custom";
-                                              isoDateTime: string;
-                                              timezone: string;
-                                            };
-                                      });
-                                    setTrigger({
-                                      ...trigger,
-                                      config: { ...base, intervalMinutes: val },
-                                    });
-                                  }}
-                                  error={
-                                    validation.fieldErrors.intervalMinutes?.[0]
-                                  }
-                                />
-                              </div>
-                              <div>
-                                <Input
-                                  id="every-timezone"
-                                  type="select"
-                                  label="Timezone"
-                                  value={
-                                    getEveryNMinutesConfig(trigger)?.timezone ||
-                                    moment.tz.guess()
-                                  }
-                                  onChange={(e) => {
-                                    const base =
-                                      getEveryNMinutesConfig(trigger) ||
-                                      ({
-                                        intervalMinutes: 15,
-                                        timezone: moment.tz.guess(),
-                                      } as {
-                                        intervalMinutes: number;
-                                        timezone: string;
-                                        anchor?:
-                                          | { kind: "topOfHour" }
-                                          | {
-                                              kind: "custom";
-                                              isoDateTime: string;
-                                              timezone: string;
-                                            };
-                                      });
-                                    setTrigger({
-                                      ...trigger!,
-                                      config: {
-                                        ...base,
-                                        timezone: e.target.value,
-                                      },
-                                    });
-                                  }}
-                                  disabled={trigger?.enabled === false}
-                                  error={validation.fieldErrors.timezone?.[0]}
-                                >
-                                  {timezoneOptions.map((tz) => (
-                                    <option key={tz.value} value={tz.value}>
-                                      {tz.label}
-                                    </option>
-                                  ))}
-                                </Input>
-                              </div>
-                              <div>
-                                <Input
-                                  id="every-anchor"
-                                  type="select"
-                                  label="Anchor"
-                                  value={(() => {
-                                    const a =
-                                      getEveryNMinutesConfig(trigger)?.anchor;
-                                    return a?.kind === "custom"
-                                      ? "custom"
-                                      : "topOfHour";
-                                  })()}
-                                  onChange={(e) => {
-                                    const kind =
-                                      e.target.value === "custom"
-                                        ? "custom"
-                                        : "topOfHour";
-                                    const existing =
-                                      getEveryNMinutesConfig(trigger) ||
-                                      ({
-                                        intervalMinutes: 15,
-                                        timezone: "UTC",
-                                      } as {
-                                        intervalMinutes: number;
-                                        timezone: string;
-                                        anchor?:
-                                          | { kind: "topOfHour" }
-                                          | {
-                                              kind: "custom";
-                                              isoDateTime: string;
-                                              timezone: string;
-                                            };
-                                      });
-                                    const nextAnchor =
-                                      kind === "topOfHour"
-                                        ? { kind: "topOfHour" as const }
-                                        : ({
-                                            kind: "custom",
-                                            isoDateTime:
-                                              new Date().toISOString(),
-                                            timezone: existing.timezone,
-                                          } as const);
-                                    setTrigger({
-                                      ...trigger,
-                                      config: {
-                                        ...existing,
-                                        anchor: nextAnchor,
-                                      },
-                                    });
-                                  }}
-                                >
-                                  <option value="topOfHour">Top of hour</option>
-                                  <option value="custom">Custom</option>
-                                </Input>
-                              </div>
-                            </div>
-
-                            {getEveryNMinutesConfig(trigger)?.anchor?.kind ===
-                              "custom" && (
-                              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                                <div>
-                                  <Input
-                                    id="every-custom-iso"
-                                    type="datetime-local"
-                                    label="Custom Anchor"
-                                    value={(() => {
-                                      const a =
-                                        getEveryNMinutesConfig(trigger)?.anchor;
-                                      if (
-                                        !a ||
-                                        a.kind !== "custom" ||
-                                        !a.isoDateTime
-                                      )
-                                        return "";
-                                      try {
-                                        return moment(a.isoDateTime)
-                                          .tz(a.timezone)
-                                          .format("YYYY-MM-DDTHH:mm");
-                                      } catch {
-                                        return "";
-                                      }
-                                    })()}
-                                    onChange={(e) => {
-                                      const existing =
-                                        getEveryNMinutesConfig(trigger) ||
-                                        ({
-                                          intervalMinutes: 15,
-                                          timezone: "UTC",
-                                        } as {
-                                          intervalMinutes: number;
-                                          timezone: string;
-                                          anchor?:
-                                            | { kind: "topOfHour" }
-                                            | {
-                                                kind: "custom";
-                                                isoDateTime: string;
-                                                timezone: string;
-                                              };
-                                        });
-                                      const a = existing.anchor;
-                                      if (!a || a.kind !== "custom") return;
-                                      const tz =
-                                        a.timezone || moment.tz.guess();
-                                      const iso = e.target.value
-                                        ? moment
-                                            .tz(
-                                              e.target.value,
-                                              "YYYY-MM-DDTHH:mm",
-                                              tz,
-                                            )
-                                            .toISOString()
-                                        : new Date().toISOString();
-                                      setTrigger({
-                                        ...trigger,
-                                        config: {
-                                          ...existing,
-                                          anchor: { ...a, isoDateTime: iso },
-                                        },
-                                      });
-                                    }}
-                                    error={
-                                      validation.fieldErrors[
-                                        "anchor.isoDateTime"
-                                      ]?.[0]
-                                    }
-                                  />
-                                </div>
-                                <div>
-                                  <Input
-                                    id="every-custom-tz"
-                                    type="select"
-                                    label="Anchor Timezone"
-                                    value={(() => {
-                                      const a =
-                                        getEveryNMinutesConfig(trigger)?.anchor;
-                                      return a && a.kind === "custom"
-                                        ? a.timezone
-                                        : moment.tz.guess();
-                                    })()}
-                                    onChange={(e) => {
-                                      const existing =
-                                        getEveryNMinutesConfig(trigger) ||
-                                        ({
-                                          intervalMinutes: 15,
-                                          timezone: moment.tz.guess(),
-                                        } as {
-                                          intervalMinutes: number;
-                                          timezone: string;
-                                          anchor?:
-                                            | { kind: "topOfHour" }
-                                            | {
-                                                kind: "custom";
-                                                isoDateTime: string;
-                                                timezone: string;
-                                              };
-                                        });
-                                      const a = existing.anchor;
-                                      if (!a || a.kind !== "custom") return;
-                                      const newTimezone = e.target.value;
-                                      const oldTimezone = a.timezone;
-                                      let newIsoDateTime = a.isoDateTime;
-                                      if (
-                                        a.isoDateTime &&
-                                        oldTimezone &&
-                                        newTimezone !== oldTimezone
-                                      ) {
-                                        try {
-                                          const wallClockTime = moment(
-                                            a.isoDateTime,
-                                          ).tz(oldTimezone);
-                                          newIsoDateTime = moment
-                                            .tz(
-                                              wallClockTime.format(
-                                                "YYYY-MM-DDTHH:mm:ss",
-                                              ),
-                                              "YYYY-MM-DDTHH:mm:ss",
-                                              newTimezone,
-                                            )
-                                            .toISOString();
-                                        } catch {
-                                          newIsoDateTime = a.isoDateTime;
-                                        }
-                                      }
-                                      setTrigger({
-                                        ...trigger!,
-                                        config: {
-                                          ...existing,
-                                          anchor: {
-                                            ...a,
-                                            timezone: newTimezone,
-                                            isoDateTime: newIsoDateTime,
-                                          },
-                                        },
-                                      });
-                                    }}
-                                    disabled={trigger?.enabled === false}
-                                    error={
-                                      validation.fieldErrors[
-                                        "anchor.timezone"
-                                      ]?.[0]
-                                    }
-                                  >
-                                    {timezoneOptions.map((tz) => (
-                                      <option key={tz.value} value={tz.value}>
-                                        {tz.label}
-                                      </option>
-                                    ))}
-                                  </Input>
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                        )}
-
-                        {trigger?.type === "time.minuteOfHour" && (
-                          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                              <Input
-                                id="minute-minute"
-                                type="number"
-                                label="Minute (0-59)"
-                                min={0}
-                                max={59}
-                                value={
-                                  getMinuteOfHourConfig(trigger)?.minute ?? 0
-                                }
-                                onChange={(e) => {
-                                  const val = Math.max(
-                                    0,
-                                    Math.min(59, Number(e.target.value) || 0),
-                                  );
-                                  const base =
-                                    getMinuteOfHourConfig(trigger) ||
-                                    ({ minute: 0, timezone: "UTC" } as {
-                                      minute: number;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger,
-                                    config: { ...base, minute: val },
-                                  });
-                                }}
-                                error={validation.fieldErrors.minute?.[0]}
-                              />
-                            </div>
-                            <div>
-                              <Input
-                                id="minute-timezone"
-                                type="select"
-                                label="Timezone"
-                                value={
-                                  getMinuteOfHourConfig(trigger)?.timezone ||
-                                  moment.tz.guess()
-                                }
-                                onChange={(e) => {
-                                  const base =
-                                    getMinuteOfHourConfig(trigger) ||
-                                    ({
-                                      minute: 0,
-                                      timezone: moment.tz.guess(),
-                                    } as {
-                                      minute: number;
-                                      timezone: string;
-                                    });
-                                  setTrigger({
-                                    ...trigger!,
-                                    config: {
-                                      ...base,
-                                      timezone: e.target.value,
-                                    },
-                                  });
-                                }}
-                                disabled={trigger?.enabled === false}
-                                error={validation.fieldErrors.timezone?.[0]}
-                              >
-                                {timezoneOptions.map((tz) => (
-                                  <option key={tz.value} value={tz.value}>
-                                    {tz.label}
-                                  </option>
-                                ))}
-                              </Input>
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    </CardContent>
-                  </Card>
+                  <TimeTriggerConfig
+                    trigger={trigger}
+                    setTrigger={setTrigger}
+                    validation={validation}
+                  />
                 )}
 
                 {/* Randomization Buckets */}
@@ -2753,60 +720,7 @@ const SpawnEditorWorkspace: React.FC = memo(() => {
                   onChange={setBucketsDraft}
                 />
 
-                <Card>
-                  <CardHeader>
-                    <div className="flex items-center justify-between">
-                      <CardTitle className="text-base font-semibold text-[rgb(var(--color-fg))]">
-                        Metadata
-                      </CardTitle>
-                      <Button
-                        type="button"
-                        variant="outline"
-                        size="sm"
-                        onClick={() => setShowMetadata((v) => !v)}
-                        aria-label="Toggle metadata"
-                      >
-                        {showMetadata ? "Hide" : "Show"}
-                      </Button>
-                    </div>
-                  </CardHeader>
-                  <CardContent>
-                    {showMetadata && (
-                      <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                        <div>
-                          <Input
-                            id="spawn-id"
-                            type="text"
-                            label="ID"
-                            value={selectedSpawn.id}
-                            disabled
-                          />
-                        </div>
-                        <div>
-                          <Input
-                            id="spawn-modified"
-                            type="text"
-                            label="Last Modified"
-                            value={formatDate(selectedSpawn.lastModified)}
-                            disabled
-                          />
-                        </div>
-
-                        <div>
-                          <Input
-                            id="spawn-assets"
-                            type="text"
-                            label="Assets"
-                            value={`${selectedSpawn.assets.length} item${
-                              selectedSpawn.assets.length === 1 ? "" : "s"
-                            }`}
-                            disabled
-                          />
-                        </div>
-                      </div>
-                    )}
-                  </CardContent>
-                </Card>
+                <MetadataCard spawn={selectedSpawn} />
               </div>
             ) : (
               <div className="text-[rgb(var(--color-muted-foreground))] text-sm">
