+++ src/components/configuration/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useAssetDraftCache } from "./useAssetDraftCache";
+export type { AssetDraftCacheEntry } from "./useAssetDraftCache";
