+++ src/components/configuration/index.ts
@@ -1,2 +1,15 @@
 export { default as SpawnEditorWorkspace } from "./SpawnEditorWorkspace";
 export { default as AssetSettingsForm } from "./asset-settings/AssetSettingsForm";
+
+// Extracted components
+export { SpawnEditorDialogs } from "./SpawnEditorDialogs";
+export { SpawnSettingsHeader } from "./SpawnSettingsHeader";
+export { BasicDetailsCard } from "./BasicDetailsCard";
+export { TriggerTypeSelector } from "./TriggerTypeSelector";
+export { MetadataCard } from "./MetadataCard";
+
+// Trigger config components
+export * from "./triggers";
+
+// Hooks
+export * from "./hooks";
